+++ lib/mongoid/shell/commands.rb
@@ -1,3 +1,4 @@
 require 'mongoid/shell/commands/base'
 require 'mongoid/shell/commands/mongo'
 require 'mongoid/shell/commands/mongodump'
+require 'mongoid/shell/commands/mongostat'