+++ lib/mongoid/shell/commands/mongo.rb
@@ -7,7 +7,7 @@ module Mongoid
         include Mongoid::Shell::Properties::Username
         include Mongoid::Shell::Properties::Password
 
-        attr_accessor :eval
+        attr_accessor :eval, :nodb, :norc, :quiet, :ipv6
 
         def initialize(attrs = {})
           super
@@ -22,7 +22,11 @@ module Mongoid
             'db address' => :host_port_and_db,
             '--username' => :username,
             '--password' => :password,
-            '--eval' => :eval
+            '--eval' => :eval,
+            '--nodb' => :nodb,
+            '--norc' => :norc,
+            '--quiet' => :quiet,
+            '--ipv6' => :ipv6,
           })
         end